requirements.txt
django==2.2.16
djangorestframework==3.9.4
psycopg2-binary==2.8.6
dj-database-url==0.5.0
